requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
    #   watchfiles
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   ipython
    #   tqdm
    #   uvicorn
comm==0.2.2
    # via ipykernel
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
executing==2.2.0
    # via stack-data
fastapi==0.115.14
    # via daily-reminder (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.174.0
    # via daily-reminder (pyproject.toml)
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-genai
google-auth-httplib2==0.2.0
    # via
    #   daily-reminder (pyproject.toml)
    #   google-api-python-client
google-auth-oauthlib==1.2.2
    # via daily-reminder (pyproject.toml)
google-genai==1.24.0
    # via daily-reminder (pyproject.toml)
googleapis-common-protos==1.70.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   google-genai
    #   mcp
    #   openai
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
ipykernel==6.29.5
    # via daily-reminder (pyproject.toml)
ipython==9.4.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via fastapi
jiter==0.10.0
    # via openai
jsonschema==4.24.0
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mcp==1.10.1
    # via daily-reminder (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
nest-asyncio==1.6.0
    # via ipykernel
oauthlib==3.3.1
    # via requests-oauthlib
openai==1.93.0
    # via daily-reminder (pyproject.toml)
packaging==25.0
    # via ipykernel
parso==0.8.4
    # via jedi
platformdirs==4.3.8
    # via jupyter-core
prompt-toolkit==3.0.51
    # via ipython
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==7.0.0
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   google-genai
    #   mcp
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via jupyter-client
python-dotenv==1.1.1
    # via
    #   daily-reminder (pyproject.toml)
    #   mcp
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   fastapi
    #   mcp
pywin32==310
    # via jupyter-core
pyyaml==6.0.2
    # via uvicorn
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-api-core
    #   google-genai
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.8
    # via fastapi-cli
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==2.3.6
    # via mcp
stack-data==0.6.3
    # via ipython
starlette==0.46.2
    # via
    #   fastapi
    #   mcp
tenacity==8.5.0
    # via google-genai
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typer==0.16.0
    # via
    #   fastapi-cli
    #   mcp
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via
    #   daily-reminder (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   mcp
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
